# Conflicts:
#	simora/src/main/java/unipr/fshmn/simora/MvcConfig.java
#	simora/src/main/java/unipr/fshmn/simora/controllers/UserController.java
#	simora/src/main/resources/templates/login.html

+++ simora/src/main/java/unipr/fshmn/simora/db/ScheduleRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called ScheduleRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface ScheduleRepository extends CrudRepository<Schedule, Integer> {
+public interface ScheduleRepository extends CrudRepository<Schedule, Long> {
 
     List<Schedule> findBySubject(String subject);
